backend/service.py
"""
Embody Backend Service

FastAPI service that provides design system exploration via Amplifier Foundation.
"""

from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

app = FastAPI(
    title="Embody API",
    description="Amplifier-powered design system exploration",
    version="0.1.0"
)

# CORS configuration
cors_origins = os.getenv("CORS_ORIGINS", "http://localhost:3456,http://127.0.0.1:3456").split(",")
app.add_middleware(
    CORSMiddleware,
    allow_origins=cors_origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.get("/")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "embody-backend",
        "version": "0.1.0"
    }


@app.post("/api/sessions/create")
async def create_session():
    """Create new design exploration session"""
    # TODO: Implement session creation
    return {"message": "Session creation endpoint - coming soon"}


@app.post("/api/sessions/{session_id}/gather-context")
async def gather_context(session_id: str):
    """Gather designer intent context"""
    # TODO: Implement context gathering
    return {"message": "Context gathering endpoint - coming soon"}


@app.post("/api/sessions/{session_id}/generate-concepts")
async def generate_concepts(session_id: str):
    """Generate design concepts"""
    # TODO: Implement concept generation
    return {"message": "Concept generation endpoint - coming soon"}


@app.post("/api/sessions/{session_id}/refine")
async def refine_concepts(session_id: str):
    """Refine concepts based on feedback"""
    # TODO: Implement refinement
    return {"message": "Refinement endpoint - coming soon"}


@app.post("/api/sessions/{session_id}/finalize")
async def finalize_direction(session_id: str):
    """Finalize and document design direction"""
    # TODO: Implement finalization
    return {"message": "Finalization endpoint - coming soon"}


@app.get("/api/sessions/{session_id}")
async def get_session(session_id: str):
    """Get session state"""
    # TODO: Implement session retrieval
    return {"message": "Session retrieval endpoint - coming soon"}


if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", "8000"))
    host = os.getenv("HOST", "0.0.0.0")
    uvicorn.run(app, host=host, port=port)
